requirements.txt
flask==3.0.0
openai==1.30.0
python-dotenv==1.0.1
httpx==0.27.0